Add StyleCop configuration and enhance test cases for numeric conversions
- Introduced .editorconfig and stylecop.json for StyleCop settings.
- Updated Directory.Build.props to include StyleCop.Analyzers.
- Refactored MathTests and PluginsAttributesTests to use Attribute.IsDefined for attribute validation.
- Modified ToDecimalMethodsTests, ToDoubleMethodsTests, ToFloatMethodsTests, and ToStringMethodsTests to use appropriate numeric suffixes.
- Added Copilot session summary for tracking progress and next steps.

+++ Musoq.Plugins.Tests/ToDecimalMethodsTests.cs
@@ -243,7 +243,7 @@ public class ToDecimalMethodsTests : LibraryBaseBaseTests
     [TestMethod]
     public void ToDecimal_WhenULongProvided_ShouldReturnDecimal()
     {
-        var result = Library.ToDecimal((ulong)123456789012);
+        var result = Library.ToDecimal(123456789012UL);
 
         Assert.AreEqual(123456789012m, result);
     }
